add function to timestamp range
